requirements.txt
scipy==1.9.3
numpy==1.23.4
h5py==3.7.0
beautifulsoup4==4.11.1
gym==0.26.2
mpi4py==3.1.3=py39h32b9844_1
PyYAML==6.0
torch==1.13.0
cython<3
mujoco-py<2.2,>=2.0
num2words==0.5.12
lxml==4.9.1
FrEIA==0.2
opencv-python==4.6.0.66
appdirs
wandb
flatten-dict
pandas
holoviews
matplotlib
pyabc[autograd] 
POT
dm_control
envpool
git+https://github.com/jacr13/Metaworld.git@fix-box-warning